Add shared wasm function for package version readout
