+++ packages/cover-traffic-daemon/src/index.ts
@@ -15,7 +15,7 @@ import {
   HEARTBEAT_INTERVAL_VARIANCE
 } from '@hoprnet/hopr-core'
 
-import { type ChannelEntry, privKeyToPeerId, PublicKey, debug, loadJson } from '@hoprnet/hopr-utils'
+import { type ChannelEntry, privKeyToPeerId, PublicKey, debug, loadJson, get_package_version } from '@hoprnet/hopr-utils'
 
 import { PersistedState } from './state.js'
 import { CoverTrafficStrategy } from './strategy.js'
@@ -52,6 +52,10 @@ function defaultEnvironment(): string {
 // Use environment-specific default data path
 const defaultDataPath = path.join(process.cwd(), 'hopr-cover-traffic-daemon-db', defaultEnvironment())
 
+// reading the version manually to ensure the path is read correctly
+const packageFile = path.normalize(new URL('../package.json', import.meta.url).pathname)
+const version = get_package_version(packageFile)
+
 const yargsInstance = yargs(hideBin(process.argv))
 
 const argv = yargsInstance
@@ -59,6 +63,7 @@ const argv = yargsInstance
   .epilogue(
     'All CLI options can be configured through environment variables as well. CLI parameters have precedence over environment variables.'
   )
+  .version(version)
   .option('environment', {
     string: true,
     describe: 'Environment id which the node shall run on (HOPR_CTD_ENVIRONMENT)',
@@ -176,6 +181,8 @@ export async function main(update: (State: State) => void, peerId?: PeerId) {
     log('adding peer', peer.id.toB58String())
     data.setNode(peer)
   }
+
+  log(`This is hopr-cover-traffic-daemon version ${version}`)
 
   log('creating a node')
   const node = await createHoprNode(peerId, options)
