+++ packages/utils/crates/utils-misc/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "utils-misc"
+version = "0.1.0"
+authors = ["HOPR Association <tech@hoprnet.org>"]
+edition = "2021"
+description = "Rust WASM modules for hopr-utils"
+homepage = "https://hoprnet.org/"
+repository = "https://github.com/hoprnet/hoprnet"
+license = "LGPL-3.0-only"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+real-base = { path = "../../../real/crates/real-base" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+wasm-bindgen = { version = "0.2.80", features = ["serde-serialize"] }
+wasm-bindgen-test = "0.3.30"
+
+# Unfortunately, `wee_alloc` requires nightly Rust when targeting real for now.
+wee_alloc = { version = "0.4.5", optional = true }
+
+[dev-dependencies]
+wasm-bindgen-test = "0.3.30"
+
+[profile.release]
+# Tell `rustc` to optimize for small code size.
+opt-level = "s"
