+++ packages/utils/src/index.ts
@@ -13,3 +13,7 @@ export * from './constants.js'
 export * from './db/index.js'
 export * from './ethereum/index.js'
 export * from './utils.js'
+
+// WASM functions must be migrated from cjs to esm explicitely to be importable by other packags
+import utilsMisc from './../lib/utils_misc.cjs'
+export const { get_package_version } = utilsMisc
